Refactor driven adapter Kubescape
