+++ example_remote_config_local/engine_core/ConfigTool.json
@@ -26,7 +26,7 @@
     },
     "ENGINE_IAC": {
         "ENABLED": "true",
-        "TOOL": "CHECKOV"
+        "TOOL": "CHECKOV|KUBESCAPE"
     },
     "ENGINE_CONTAINER": {
         "ENABLED": "true",
