+++ tools/devsecops_engine_tools/engine_sast/engine_iac/src/infrastructure/driven_adapters/kubescape/kubescape_deserealizator.py
@@ -28,3 +28,48 @@ class KubescapeDeserealizator:
             list_open_findings.append(finding_open)
 
         return list_open_findings
+
+    def extract_failed_controls(self, data):
+        result_extracted_data = []
+        results = data.get("results", [])
+        resources = {resource.get("resourceID"): resource for resource in data.get("resources", [])}
+        frameworks = data.get("summaryDetails", {}).get("frameworks", [])
+
+        for result in results:
+            resource_id = result.get("resourceID")
+            controls = result.get("controls", [])
+
+            for control in controls:
+                if control.get("status", {}).get("status") == "failed":
+                    control_id = control.get("controlID")
+                    name = control.get("name")
+                    resource = resources.get(resource_id)
+
+                    if resource:
+                        relative_path = resource.get("source", {}).get("relativePath", "").replace("\\", "/")
+                        severity_score = self.get_severity_score(frameworks, control_id)
+
+                        result_extracted_data.append({
+                            "id": control_id,
+                            "description": name,
+                            "where": relative_path,
+                            "severity": severity_score
+                        })
+
+        return result_extracted_data
+
+    def get_severity_score(self, frameworks, control_id):
+        classifications = {
+            (0.0, 0.0): "none",
+            (0.1, 3.9): "low",
+            (4.0, 6.9): "medium",
+            (7.0, 8.9): "high",
+            (9.0, 10.0): "critical"
+        }
+        for framework in frameworks:
+            control_object = framework.get("controls", {}).get(control_id, {})
+            if control_object:
+                for range_tuple, classification in classifications.items():
+                    if range_tuple[0] <= control_object.get("scoreFactor", 0.0) <= range_tuple[1]:
+                        return classification
+        return None
