+++ tools/devsecops_engine_tools/engine_sast/engine_iac/src/infrastructure/driven_adapters/kubescape/kubescape_tool.py
@@ -3,6 +3,8 @@ import subprocess
 import platform
 import requests
 import distro
+import os
+import time
 from devsecops_engine_tools.engine_sast.engine_iac.src.domain.model.gateways.tool_gateway import (
     ToolGateway,
 )
@@ -11,9 +13,6 @@ from devsecops_engine_tools.engine_sast.engine_iac.src.domain.model.config_tool 
 )
 from devsecops_engine_tools.engine_sast.engine_iac.src.infrastructure.driven_adapters.kubescape.kubescape_deserealizator import (
     KubescapeDeserealizator,
-)
-from devsecops_engine_tools.engine_sast.engine_iac.src.infrastructure.helpers.file_generator_tool import (
-    generate_file_from_kubescape,
 )
 from devsecops_engine_tools.engine_utilities.utils.logger_info import MyLogger
 from devsecops_engine_tools.engine_utilities import settings
@@ -59,67 +58,29 @@ class KubescapeTool(ToolGateway):
             except Exception as e:
                 logger.error(f"Error installing Kubescape: {e}")
 
-    def execute_kubescape(self, folders_to_scan, prefix):
+    def execute_kubescape(self, folders_to_scan, prefix, platform_to_scan):
+        outputs_json = []
         for folder in folders_to_scan:
-            command = [prefix, "scan", "framework", "nsa", folder, "--format", "json", "--format-version", "v2", "--output", "results_kubescape.json", "-v"]
-            try:
-                subprocess.run(command, capture_output=True)
-            except subprocess.CalledProcessError as e:
-                logger.error(f"Error during Kubescape execution: {e}")
+            if "k8s" in platform_to_scan:
+                timestamp = time.strftime("%Y%m%d-%H%M%S")
+                output_filename = f"results_kubescape_{timestamp}.json"
+                command = [prefix, "scan", "framework", "nsa", folder, "--format", "json", "--format-version", "v2",
+                           "--output", output_filename, "-v"]
+                try:
+                    subprocess.run(command, capture_output=True)
+                    outputs_json.append(output_filename)
+                except subprocess.CalledProcessError as e:
+                    logger.error(f"Error during Kubescape execution: {e}")
+        return outputs_json
 
-    def load_json(self):
+    def load_json(self, json_name):
         try:
-            with open("results_kubescape.json") as file:
+            with open(json_name) as file:
                 return json.load(file)
         except FileNotFoundError:
-            logger.error("The file results_kubescape.json does not exist.")
+            logger.error(f"The file {json_name} does not exist.")
         except json.JSONDecodeError:
             logger.error("The JSON result is empty.")
-        return None
-
-    def extract_failed_controls(self, data):
-        result_extracted_data = []
-        results = data.get("results", [])
-        resources = {resource.get("resourceID"): resource for resource in data.get("resources", [])}
-        frameworks = data.get("summaryDetails", {}).get("frameworks", [])
-
-        for result in results:
-            resource_id = result.get("resourceID")
-            controls = result.get("controls", [])
-
-            for control in controls:
-                if control.get("status", {}).get("status") == "failed":
-                    control_id = control.get("controlID")
-                    name = control.get("name")
-                    resource = resources.get(resource_id)
-
-                    if resource:
-                        relative_path = resource.get("source", {}).get("relativePath", "").replace("\\", "/")
-                        severity_score = self.get_severity_score(frameworks, control_id)
-
-                        result_extracted_data.append({
-                            "id": control_id,
-                            "description": name,
-                            "where": relative_path,
-                            "severity": severity_score
-                        })
-
-        return result_extracted_data
-
-    def get_severity_score(self, frameworks, control_id):
-        classifications = {
-            (0.0, 0.0): "none",
-            (0.1, 3.9): "low",
-            (4.0, 6.9): "medium",
-            (7.0, 8.9): "high",
-            (9.0, 10.0): "critical"
-        }
-        for framework in frameworks:
-            control_object = framework.get("controls", {}).get(control_id, {})
-            if control_object:
-                for range_tuple, classification in classifications.items():
-                    if range_tuple[0] <= control_object.get("scoreFactor", 0.0) <= range_tuple[1]:
-                        return classification
         return None
 
     def run_tool(self, config_tool: ConfigTool, folders_to_scan, environment, platform_to_scan, secret_tool):
@@ -142,23 +103,35 @@ class KubescapeTool(ToolGateway):
             elif os_platform == "Windows":
                 file = "kubescape-windows-latest.exe"
                 self.install_tool_windows(file, base_url + file)
-                command_prefix = "./kubescape-windows-latest"
+                command_prefix = f"./{file}"
+            elif os_platform == "Darwin":
+                file = "kubescape-macos-latest"
+                self.install_tool(file, base_url + file)
+                command_prefix = f"./{file}"
             else:
                 logger.warning(f"{os_platform} is not supported.")
-                return None
-
-            self.execute_kubescape(folders_to_scan, command_prefix)
-            data = self.load_json()
-
-            if not data:
                 return [], None
-            else:
-                result_extracted_data = self.extract_failed_controls(data)
-                kubescape_deserealizator = KubescapeDeserealizator()
-                finding_list = kubescape_deserealizator.get_list_finding(result_extracted_data)
-                path_file_results = generate_file_from_kubescape(data)
 
-                return finding_list, path_file_results
+            outputs_json = self.execute_kubescape(folders_to_scan, command_prefix, platform_to_scan)
+
+            total_findings = []
+            path_json_result = []
+            for json_name in outputs_json:
+
+                path = os.path.abspath(json_name)
+                path_json_result.append(path)
+
+                data = self.load_json(json_name)
+
+                if not data:
+                    return [], None
+                else:
+                    kubescape_deserealizator = KubescapeDeserealizator()
+                    result_extracted_data = kubescape_deserealizator.extract_failed_controls(data)
+                    finding_list = kubescape_deserealizator.get_list_finding(result_extracted_data)
+                    total_findings.extend(finding_list)
+
+            return total_findings, path_json_result
 
         else:
             return [], None
