+++ tools/devsecops_engine_tools/engine_sast/engine_iac/src/infrastructure/helpers/file_generator_tool.py
@@ -60,49 +60,6 @@ def generate_file_from_tool(tool, result_list, rules_doc):
             logger.error(f"Error during handling checkov json integrator {ex}")
 
 
-def generate_file_from_kubescape(result_list):
-    try:
-        if result_list:
-            all_failed_checks = []
-            summary_passed = 0
-            summary_failed = 0
-            kubescape_version = result_list.get("metadata", {}).get("scanMetadata", {}).get("kubescapeVersion")
-
-            frameworks = result_list.get("summaryDetails", {}).get("frameworks", [])
-
-            for framework in frameworks:
-                controls = framework.get("controls", {})
-                for control_id, control_data in controls.items():
-                    if control_data['status'] == 'failed':
-                        all_failed_checks.append({control_id: control_data})
-                        summary_failed += 1
-
-                    if control_data['status'] == 'passed':
-                        summary_passed += 1
-
-            file_name = "results_scan.json"
-            results_data = {
-                "check_type": "Dockerfile, Kubernetes and CloudFormation",
-                "results": {
-                    "failed_checks": all_failed_checks,
-                },
-                "summary": {
-                    "passed": summary_passed,
-                    "failed": summary_failed,
-                    "kubescape_version": kubescape_version,
-                },
-            }
-
-            with open(file_name, "w") as json_file:
-                json.dump(results_data, json_file, indent=4)
-
-            absolute_path = os.path.abspath(file_name)
-            return absolute_path
-
-    except Exception as ex:
-        logger.error(f"Error during handling kubescape json integrator {ex}")
-
-
 def update_fields(check_result, rules_doc):
     rule_info = rules_doc.get(check_result.get("check_id"), {})
